MFC: The ai_addrlen of a struct addrinfo used to be a size_t, per RFC 2553.  In XNS5.2, and subsequently in POSIX-2001 and RFC 3493, it was changed to a socklen_t.  And, the n_net of a struct netent used to be an unsigned long integer.  In XNS5, and subsequently in POSIX-2001, it was changed to an uint32_t.  To accomodate for this while preserving ABI compatibility with the old interface, we need to prepend or append 32 bits of padding, depending on the (LP64) architecture's endianness.
	include/netdb.h:		1.36 (partly)
	lib/libc/net/getaddrinfo.c:	1.69
	lib/libc/net/getnetbydns.c:	1.30
	lib/libc/net/getnetbyht.c:	1.15
	lib/libc/net/getnetbynis.c:	1.18
	lib/libc/net/getnetnamadr.c:	1.20 (partly)
	lib/libc/net/netdb_private.h:	1.9

+++ lib/libc/net/getnetnamadr.c
@@ -120,7 +120,8 @@ getnetbyname_r(const char *name, struct netent *ne, struct netent_data *ned)
 }
 
 int
-getnetbyaddr_r(u_long addr, int af, struct netent *ne, struct netent_data *ned)
+getnetbyaddr_r(uint32_t addr, int af, struct netent *ne,
+    struct netent_data *ned)
 {
 	int rval;
 
@@ -170,7 +171,7 @@ getnetbyaddr(u_long addr, int af)
 
 	if ((nd = __netdata_init()) == NULL)
 		return NULL;
-	if (getnetbyaddr_r(addr, af, &nd->net, &nd->data) != 0)
+	if (getnetbyaddr_r((uint32_t)addr, af, &nd->net, &nd->data) != 0)
 		return NULL;
 	return &nd->net;
 }
